♻️ Refactor firebase location
