+++ functions/modules/firebase.js
@@ -2,6 +2,21 @@ const admin = require('firebase-admin');
 
 admin.initializeApp();
 
-export const db = admin.firestore();
+const database = admin.firestore();
+const db = database;
+const { FieldValue } = admin.firestore;
 
-export default db;
+const getCollectionIfDefined = root =>
+  root ? database.collection(root) : null;
+
+const userCollection = getCollectionIfDefined(process.env.FIRESTORE_USER_ROOT);
+const nftMintSubscriptionCollection = getCollectionIfDefined(
+  process.env.FIRESTORE_NFT_MINT_SUBSCRIPTION_ROOT
+);
+
+module.exports = {
+  db,
+  FieldValue,
+  userCollection,
+  nftMintSubscriptionCollection,
+};
