+++ src/server/api/index.js
@@ -4,7 +4,7 @@ const session = require('express-session');
 const bodyParser = require('body-parser');
 const cookieParser = require('cookie-parser');
 const FirestoreStore = require('firestore-store')(session);
-const { db } = require('./util/firebase');
+const { db } = require('../modules/firebase');
 const users = require('./routes/users');
 const civic = require('./routes/civic');
 const reader = require('./routes/reader');
