+++ src/server/modules/firebase.js
@@ -2,7 +2,7 @@ const admin = require('firebase-admin');
 const {
   FIRESTORE_USER_ROOT,
   FIRESTORE_NFT_MINT_SUBSCRIPTION_ROOT,
-} = require('../../config/config');
+} = require('../config/config');
 
 let database;
 if (!process.env.CI) {
@@ -10,7 +10,7 @@ if (!process.env.CI) {
     admin.initializeApp();
   } else {
     /* eslint-disable-next-line global-require */
-    const serviceAccount = require('../../config/serviceAccountKey.json');
+    const serviceAccount = require('../config/serviceAccountKey.json');
     admin.initializeApp({
       credential: admin.credential.cert(serviceAccount),
     });
